+++ plugins/doc_fragments/powerscale.py
@@ -1,22 +1,13 @@
 # -*- coding: utf-8 -*-
-# Copyright: (c) 2019, Dell EMC.
+# Copyright: (c) 2019, Dell Technologies
 
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 
 class ModuleDocFragment(object):
-
-    DOCUMENTATION = r'''
-options:
-  - See respective platform section for more details
-requirements:
-  - See respective platform section for more details
-notes:
-  - Ansible modules are available for Dell EMC PowerScale Storage Platform
-'''
     # Documentation fragment for PowerScale (powerscale)
-    POWERSCALE = r'''
+    DOCUMENTATION = r'''
     options:
         onefs_host:
             description:
@@ -52,9 +43,8 @@ notes:
             - the password of the PowerScale cluster.
             required: True
     requirements:
-      - A Dell EMC PowerScale Storage system.
-      - Ansible 2.10, 2.11 or 2.12.
+      - A Dell PowerScale Storage system. Ansible 2.11, 2.12 or 2.13.
     notes:
       - The modules present in this collection named as 'dellemc.powerscale'
-        are built to support the Dell EMC PowerScale storage platform.
+        are built to support the Dell PowerScale storage platform.
     '''