+++ plugins/modules/group.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright: (c) 2019, DellEMC
+# Copyright: (c) 2019, Dell Technologies
 
 # Apache License version 2.0 (see MODULE-LICENSE or http://www.apache.org/licenses/LICENSE-2.0.txt)
 
@@ -21,7 +21,7 @@ description:
   delete group,  get group, add users and remove users.
 
 extends_documentation_fragment:
-  - dellemc.powerscale.dellemc_powerscale.powerscale
+  - dellemc.powerscale.powerscale
 author:
 - P Srinivas Rao (@srinivas-rao5) <ansible.team@dell.com>
 options:
@@ -217,7 +217,7 @@ group_details:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.dellemc.powerscale.plugins.module_utils.storage.dell \
-    import dellemc_ansible_powerscale_utils as utils
+    import utils
 import re
 
 LOG = utils.get_logger('group')